+++ statistics/StatManager.cpp
@@ -8,21 +8,31 @@ float StatManager::getOverallStat(STATISTIC stat) {
     return _STATS_OVERALL[stat];
 }
 
-float StatManager::getStatThisSave(STATISTIC stat) {
-    return _STATS_THIS_SAVE[stat];
+float StatManager::getStatThisRun(STATISTIC stat) {
+    return _STATS_THIS_RUN[stat];
 }
 
 void StatManager::increaseStat(STATISTIC stat, float amt) {
     if (DISABLE_STATS) return;
 
     _STATS_OVERALL[stat] += amt;
-    _STATS_THIS_SAVE[stat] += amt;
+    _STATS_THIS_RUN[stat] += amt;
 
-    AchievementManager::checkAchievementsOnStatIncrease(stat, getStatThisSave(stat));
+    AchievementManager::checkAchievementsOnStatIncrease(stat, getStatThisRun(stat));
 }
 
 void StatManager::loadOverallStats() {
-    std::string path = getLocalLowPath() + "\\stats.config";
+    if (SELECTED_SAVE_FILE == SAVE_FILE_NOT_SELECTED) {
+        MessageManager::displayMessage("Tried to load stats before save file was selected", 5, WARN);
+        return;
+    }
+
+    std::string path = getLocalLowPath() + "\\save" + std::to_string(SELECTED_SAVE_FILE);
+    if (!std::filesystem::is_directory(path + "\\")) {
+        std::filesystem::create_directory(path);
+    }
+
+    path += "\\stats.config";
     std::ifstream in(path);
 
     if (!in.good()) {
@@ -48,7 +58,17 @@ void StatManager::loadOverallStats() {
 void StatManager::saveOverallStats() {
     if (DISABLE_STATS) return;
 
-    std::string path = getLocalLowPath() + "\\stats.config";
+    if (SELECTED_SAVE_FILE == SAVE_FILE_NOT_SELECTED) {
+        MessageManager::displayMessage("Tried to save stats before save file was selected", 5, WARN);
+        return;
+    }
+
+    std::string path = getLocalLowPath() + "\\save" + std::to_string(SELECTED_SAVE_FILE);
+    if (!std::filesystem::is_directory(path + "\\")) {
+        std::filesystem::create_directory(path);
+    }
+
+    path += "\\stats.config";
 
     try {
         if (!std::filesystem::remove(path)) {
@@ -69,10 +89,10 @@ void StatManager::saveOverallStats() {
     }
 }
 
-void StatManager::setStatThisSave(STATISTIC stat, float val) {
+void StatManager::setStatThisRun(STATISTIC stat, float val) {
     if (DISABLE_STATS) return;
 
-    _STATS_THIS_SAVE[stat] = val;
+    _STATS_THIS_RUN[stat] = val;
 }
 
 void StatManager::setOverallStat(STATISTIC stat, float val) {
@@ -81,8 +101,14 @@ void StatManager::setOverallStat(STATISTIC stat, float val) {
     _STATS_OVERALL[stat] = val;
 }
 
-void StatManager::resetStatsForThisSave() {
+void StatManager::resetStatsForThisRun() {
     for (int i = 0; i < NUM_STATS; i++) {
-        setStatThisSave((STATISTIC)i, 0);
+        setStatThisRun((STATISTIC)i, 0);
+    }
+}
+
+void StatManager::resetOverallStats() {
+    for (int i = 0; i < NUM_STATS; i++) {
+        setOverallStat((STATISTIC)i, 0);
     }
 }
